pyproject.toml
[project]
name = "lyricsflow"
version = "0.1.0"
description = "Lyrics aggregation middleware"
readme = "Readme.md"
requires-python = ">=3.9"
license = {text = "Apache-2.0"}
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "fastapi",
    "uvicorn",
    "httpx",
    "pycryptodome",
    "rapidfuzz",
    "beautifulsoup4",
    "pydantic-settings",
    "python-multipart",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
